Add files via upload
For this project, R software and QGIS were used to create visualizations and a map showing the number of deaths that occurred in care homes in Wales, United Kingdom since the start of the COVID-19 pandemic. Take a look at the "Presentation_poster.pdf" to get more information on the source of the data and how it was processed.